address review comments and adds ReadOnlyTokenStore
