+++ hedera-node/src/main/java/com/hedera/services/txns/crypto/validators/AdjustAllowanceChecks.java
@@ -25,7 +25,7 @@ import com.hedera.services.exceptions.InvalidTransactionException;
 import com.hedera.services.state.enums.TokenSupplyType;
 import com.hedera.services.state.merkle.MerkleUniqueToken;
 import com.hedera.services.state.submerkle.FcTokenAllowanceId;
-import com.hedera.services.store.TypedTokenStore;
+import com.hedera.services.store.ReadOnlyTokenStore;
 import com.hedera.services.store.models.Account;
 import com.hedera.services.store.models.Id;
 import com.hedera.services.store.models.NftId;
@@ -57,7 +57,7 @@ public class AdjustAllowanceChecks extends AllowanceChecks {
 	}
 
 	@Override
-	ResponseCodeEnum validateTokenAmount(final Account ownerAccount,
+	public ResponseCodeEnum validateTokenAmount(final Account ownerAccount,
 			final long amount,
 			final Token token,
 			final Id spender) {
@@ -78,11 +78,11 @@ public class AdjustAllowanceChecks extends AllowanceChecks {
 	}
 
 	@Override
-	ResponseCodeEnum validateSerialNums(
+	public ResponseCodeEnum validateSerialNums(
 			final List<Long> serialNums,
 			final Account ownerAccount,
 			final Token token,
-			final TypedTokenStore tokenStore,
+			final ReadOnlyTokenStore tokenStore,
 			final Id spender) {
 		if (hasRepeatedSerials(serialNums)) {
 			return REPEATED_SERIAL_NUMS_IN_NFT_ALLOWANCES;
@@ -124,7 +124,7 @@ public class AdjustAllowanceChecks extends AllowanceChecks {
 	}
 
 	@Override
-	ResponseCodeEnum validateAmount(final long amount, final Account ownerAccount, final Id spender) {
+	public ResponseCodeEnum validateAmount(final long amount, final Account ownerAccount, final Id spender) {
 		final var existingAllowances = ownerAccount.getCryptoAllowances();
 		final var key = spender.asEntityNum();
